[FLINK-32618][jdbc-driver] Remove flink-core dependency from jdbc
