+++ flink-table/flink-sql-client/src/main/java/org/apache/flink/table/client/cli/CliChangelogResultView.java
@@ -21,10 +21,10 @@ package org.apache.flink.table.client.cli;
 import org.apache.flink.annotation.VisibleForTesting;
 import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.table.client.gateway.ResultDescriptor;
-import org.apache.flink.table.client.gateway.SqlExecutionException;
 import org.apache.flink.table.client.gateway.TypedResult;
 import org.apache.flink.table.client.gateway.result.ChangelogResult;
 import org.apache.flink.table.data.RowData;
+import org.apache.flink.table.gateway.SqlExecutionException;
 import org.apache.flink.table.utils.print.PrintStyle;
 
 import org.jline.keymap.KeyMap;
