+++ flink-table/flink-sql-client/src/main/java/org/apache/flink/table/client/cli/CliView.java
@@ -19,7 +19,7 @@
 package org.apache.flink.table.client.cli;
 
 import org.apache.flink.api.java.tuple.Tuple2;
-import org.apache.flink.table.client.gateway.SqlExecutionException;
+import org.apache.flink.table.gateway.SqlExecutionException;
 
 import org.jline.keymap.BindingReader;
 import org.jline.keymap.KeyMap;
