+++ flink-table/flink-sql-client/src/main/java/org/apache/flink/table/client/gateway/result/MaterializedCollectStreamResult.java
@@ -19,8 +19,8 @@
 package org.apache.flink.table.client.gateway.result;
 
 import org.apache.flink.annotation.VisibleForTesting;
-import org.apache.flink.table.client.gateway.StatementResult;
 import org.apache.flink.table.data.RowData;
+import org.apache.flink.table.gateway.StatementResult;
 import org.apache.flink.types.RowKind;
 import org.apache.flink.util.CollectionUtil;
