+++ flink-table/flink-sql-client/src/test/java/org/apache/flink/table/client/gateway/result/ChangelogCollectResultTest.java
@@ -23,11 +23,11 @@ import org.apache.flink.table.api.DataTypes;
 import org.apache.flink.table.api.ResultKind;
 import org.apache.flink.table.catalog.Column;
 import org.apache.flink.table.catalog.ResolvedSchema;
-import org.apache.flink.table.client.gateway.SqlExecutionException;
-import org.apache.flink.table.client.gateway.StatementResult;
 import org.apache.flink.table.client.gateway.TypedResult;
 import org.apache.flink.table.data.GenericRowData;
 import org.apache.flink.table.data.RowData;
+import org.apache.flink.table.gateway.SqlExecutionException;
+import org.apache.flink.table.gateway.StatementResult;
 import org.apache.flink.util.CloseableIterator;
 
 import org.junit.jupiter.api.Test;
