+++ flink-table/flink-sql-client/src/test/java/org/apache/flink/table/client/util/CliClientTestUtils.java
@@ -21,7 +21,7 @@ package org.apache.flink.table.client.util;
 import org.apache.flink.table.api.ResultKind;
 import org.apache.flink.table.catalog.ResolvedSchema;
 import org.apache.flink.table.client.SqlClient;
-import org.apache.flink.table.client.gateway.StatementResult;
+import org.apache.flink.table.gateway.StatementResult;
 import org.apache.flink.table.gateway.rest.SqlGatewayRestEndpointStatementITCase;
 import org.apache.flink.table.planner.functions.casting.RowDataToStringConverterImpl;
 import org.apache.flink.table.utils.DateTimeUtils;
